Fixed dashboard height error + login label
The page_height variable was never changed, the designer.page_height
was, thus when saving a dashboard the designer.scan() => computed page
height was not passed to the save method since it only used the old
value again.

This bug was reported by Paul Reed over at the forums here (illustrated
with a large blue arrow): http://openenergymonitor.org/emon/node/2013

Changed so it uses the designer.page_height variable instead.

This fix also makes the "Remember me" text a label around the checkbox
according to ui/ux web standards the label should be clickable. The
bootstrap default style was used for this. (see
http://twitter.github.io/bootstrap/base-css.html#forms)