Abstract MySQL out from the core code and add a simple abstraction layer
The current PHP code uses mysql directly everywhere. Even when using timestore as the main database mysql
still is a hard dependancy. To ease the transistion to possibly using Postgres and/or SQLite in combination
with timestore, it is neccesary to abstract out mysql into an abstraction layer and make calls generic.

This patch does just that and because of that adds Lib/db.php, the database abstraction layer. Queries which
are standard SQL are just passed to this library.

Signed-off-by: Oliver Schinagl <oliver@schinagl.nl>