Remove useless MySQL propriatary backtics
Nearly all backtics used are not required. The only one that will clash is 'convert' as that is a reserved
MySQL reserveword [0].

This patch cleans up all backtics as those are not cross-db and propriatery to MySQL. The Ansi-SQL version
of double quotes needs special support from MySQL which isn't enabled by default, so we can't use that.

By using this patch, it will become easier to use other databases next to MySQL for datastorage.

[0] http://dev.mysql.com/doc/mysqld-version-reference/en/mysqld-version-reference-reservedwords-5-0.html

Signed-off-by: Oliver Schinagl <oliver@schinagl.nl>