Add PostgreSQL support
With the previous cleanup, it is now quite easy to expand further
database support. This patch adds support for the PostgreSQL database.

There is a new concept of dealing with time as an actual timestamp or
the timeseries log. A new unused function is added that needs some
tuning and testing, got_data(), which uses the idea of using the
database's resolution for time rather then calculating our own using
date_trunc.

Signed-off-by: Oliver Schinagl <oliver@schinagl.nl>