Dashboard menu
The list of dashboard displayed on class="greydashmenu" is not easily accessible on mobile phones and flows over the grey bar when you have a lot of dashboards.If have moved the dashboard items to a drop down menu under the Dashboards menu item. There is more space for dashboard items this way, it cleans up the interface when using it on a desktop en surely makes it easier (or finally possible) to navigate dashboards on mobile devices.Just an idea, comments are welcome!