Improved button toggling
Altered the "inval" source so that when the feed is updated externally (eg another emoncms instance or a physical local on/off button) the buttons next click toggles the current state rather than the last set status. This is to correct the need to click the button twice if the status has changed since last click, despite the fact the colour has changed, so first click actually appears to do nothing.