Update PHPTimeSeries.php
In function get_data try to find the first valid data entry within the given timeframe and set $start to that time
This will 'sync' the start of the search through the dataset to the first real data entry. Preventing problems if $limitinterval = true and the passed &start   value is based on for example browser time.