Use different emonhub script in docs
The emonhub/upgrade script serves the same purpose as emonhub/install,
except that it checks to see if the old raspberrypi module is installed,
and if so removes it first.
Emonhub will not co-exist with the rpi module as they both access the
same port.
I've discussed this with Paul Burnell first -
http://openenergymonitor.org/emon/node/11947#comment-37960