Update feed_model.php
Added line 456 to mitigate divide by 0 errors;
This is *NOT* the correct way to fix this issue, this is a reference only to help some of you understand where the issue is.
I have not worked out yet where the variable $outinterval is supposed to be populated from, but at this point in the code its either null or 0 - and this un turn causes the divide by 0 errors.

My code simply detects when $outinterval is less than 1 and sets this variable to 1 before its used. (on line 457 in this version).

Applying my code appears to fix the connection issues with the android app, fixing the divide by zero cleaned that right up, but my fix may have other consequences (although quite unlikely).