Make dropdowns a maximum height and allow them to scroll
This aids usability when you have a long list of items. Moving your mouse all the way down can be tedious. Scrolling the list is easier.

In particular this is useful when you have a long list of dashboards which ends up going off the screen. Now it'll scroll so you can get to all of the items. This was picked up here:
https://community.openenergymonitor.org/t/dashboard-menu-does-not-scroll-not-all-dashboards-accessible-from-it