Modified multigraph_edit.js and /multigraph.js to enable the user to choose flot options.
mutligraph_edit.js changes:
Options for detail level - i.e. basic is the original set, advanced open out the following options
Choice of graph display type as Bars, Lines and Lines With Steps
For Bars you can then choose the display width of the bar
The Left/ Right axis choice has been changed to a Radio to remove the potential of having both unselected
Each data feed displayed now has a thicker row separator to make it between clearer each feed

I've made each of these specified to the series in case the user wants to do something interesting, or in case they
want something different between the 2 axis etc.

multigraph.js changes:
Modified it to support the additonal config items created in mutligraph_edit.js
The tooltip has been changed to show the date and to remove any trailing zeros from the value which means that Watt
values look a little prettier, but whilst retaining 2 decimal places for items liek kWh figures.
Tooltip now differentiates between a change of data series and also show the height of that segment. e.g. with a
stacked graph with 2 entries of 10 with 5 on top, it used to show the top as 15 as this is the bars height, now it
shows 5 for the highlighted segment. This doesn't work well where lines are too close, but I believe this is a
restriction with the way flot returns which series is being displayed.
Tooltip will also only show the date without time where datatype=2 (e.g. day level values)
I found the "toLocaleDateString" as a simple way to specify the date and time values and force the time to include
a leading 0, but currently have this stuck with "en-GB" as I do not know how to pick out the browsers locale setting