Prevent MQTT Client exception
Issue - subscribing to a topic when the MQTT client was not connected caused an exception; not fatal but not nice.

The onConnected callback is only called if there is a CONNACK from the server.

The onConnected callback assumed than any call to it meant a successful connection.  This is not correct, only a response of '0' means success so $connected now set as such.

The reconnect try, did not check the connected flag before attempting to subscribe to the topic.  By doing so the exception can be avoided.