Updated MQTT connection and subscription logic plus persistent connection
Modified the logic around the connection to and subscription to MQTT server to prevent exceptions if there is no connection to be made.  The subscription ID is returned in the log.

A persistent connection is made as per @pb66 code.  The ID returned now, is always the same number. Without it, the number returned increased.