imporve the way the mqtt queue gets cleared
Only processes a single item each iteration of the outer loop, that loops faster if there is a queue and gets throttled only when queue is empty.
Qos2 timestamped data provision included, but not for release, just for testing and to indicate where the code might go.
Qos settings for publish and subscribe could be different and user selectable in settings.php if that is desirable once cliebt id is sorted.