Add Git info to administration page
Add the git URL and the branch into the administration module.  When answering queries, it is often useful to know which branch the user is on as some questions may have already been solved in the master branch.

I have taken the URL from ls-remote. Not being an expert on Git this may not be the best method but was the only one I could find to just get the origin URL.