Update service-runner to python3.
Port service-runner to python3.

- Change indentation to 4 spaces to conform with the rest of the
project.

- Utilise redis' blpop method to block until there is something to do,
instead of spinning with a short sleep.

- There was a lot of code fighting against itself, like signal handlers
which didn't do any cleanup, other than printing a message, which
complicated the main loop. Since the program is running under a service
supervisor (presumably systemd) this is unnecessary, so remove it.

- Utilise the print(..., flush=True) syntax.

- Avoid subprocess.call(..., shell=True). The command lines are so
simple that we can (and already do) parse them ourselves.

- Avoid wrapping too much code in a try/except. There is only one call
to redis and therefore only one place where the connection can fail, so
only handle those errors there.

- Move everything into a main function, to keep linters happy.

Tested on emonpi 2 with raspbian buster.