Fix URL Encoding of Input API example with time to close #1621
This fixes Issue #1621 by URL encoding the Input API example
which uses a text string for the date and time. The date
format used includes the current system timezone which
contains a "+" character which, if not URL encoded, will be
interpreted as a space. This causes a failure in the time
parsing.

The link in the live example will now behave correctly when clicked.