Show disabled services on Admin page
The Admin page did not handle a situation where a service was loaded in
systemd but purposfully disabled.  Dead services were reported as errors
whether they were enabled or disabled.  Such services are now displayed
in a new 'Disabled' state with a grey status badge (as opposed to the
original red/green badges).